+++ tests/codegen/fib.mips
@@ -24,7 +24,17 @@ data_10: .asciiz "Substring out of range"
 .word 0
 data_11: .asciiz "Enter n to find nth fibonacci number!\n"
 .word 0
-data_12: .asciiz "\n"
+data_12: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_13: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_14: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_15: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_16: .asciiz "\n"
+.word 0
+data_17: .asciiz "RuntimeError: Function call in a void instance"
 
 
 types_table:
@@ -649,19 +659,33 @@ function_main_at_Main:
 	addi $sp $sp -4
 	sw $fp 0($sp)
 	addi $fp $sp 8
-	addi $sp $sp -36
+	addi $sp $sp -56
 	la $t0 data_11
-	sw $t0 -40($fp)
-	lw $t0 -40($fp)
+	sw $t0 -64($fp)
+	lw $t0 -64($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	jal __init_at_String
 	addi $sp $sp 4
-	sw $v0 -36($fp)
+	sw $v0 -60($fp)
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -56($fp)
+	lw $t0 -56($fp)
+	bne $t0 $zero label_3
+	j label_4
+	label_3:
+	li $v0 4
+	la $a0 data_12
+	syscall
+	li $v0 10
+	syscall
+	label_4:
 	lw $t0 0($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
-	lw $t0 -36($fp)
+	lw $t0 -60($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	lw $t0 0($fp)
@@ -673,7 +697,21 @@ function_main_at_Main:
 	lw $t0 24($t0)
 	jalr $ra $t0
 	addi $sp $sp 8
-	sw $v0 -44($fp)
+	sw $v0 -52($fp)
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -48($fp)
+	lw $t0 -48($fp)
+	bne $t0 $zero label_5
+	j label_6
+	label_5:
+	li $v0 4
+	la $a0 data_13
+	syscall
+	li $v0 10
+	syscall
+	label_6:
 	lw $t0 0($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
@@ -686,11 +724,25 @@ function_main_at_Main:
 	lw $t0 12($t0)
 	jalr $ra $t0
 	addi $sp $sp 4
-	sw $v0 -24($fp)
+	sw $v0 -44($fp)
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -40($fp)
+	lw $t0 -40($fp)
+	bne $t0 $zero label_7
+	j label_8
+	label_7:
+	li $v0 4
+	la $a0 data_14
+	syscall
+	li $v0 10
+	syscall
+	label_8:
 	lw $t0 0($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
-	lw $t0 -24($fp)
+	lw $t0 -44($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	lw $t0 0($fp)
@@ -702,11 +754,25 @@ function_main_at_Main:
 	lw $t0 28($t0)
 	jalr $ra $t0
 	addi $sp $sp 8
-	sw $v0 -28($fp)
+	sw $v0 -36($fp)
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -32($fp)
+	lw $t0 -32($fp)
+	bne $t0 $zero label_9
+	j label_10
+	label_9:
+	li $v0 4
+	la $a0 data_15
+	syscall
+	li $v0 10
+	syscall
+	label_10:
 	lw $t0 0($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
-	lw $t0 -28($fp)
+	lw $t0 -36($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	lw $t0 0($fp)
@@ -718,19 +784,33 @@ function_main_at_Main:
 	lw $t0 20($t0)
 	jalr $ra $t0
 	addi $sp $sp 8
-	sw $v0 -32($fp)
-	la $t0 data_12
-	sw $t0 -16($fp)
-	lw $t0 -16($fp)
+	sw $v0 -28($fp)
+	la $t0 data_16
+	sw $t0 -24($fp)
+	lw $t0 -24($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	jal __init_at_String
 	addi $sp $sp 4
-	sw $v0 -12($fp)
+	sw $v0 -20($fp)
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -16($fp)
+	lw $t0 -16($fp)
+	bne $t0 $zero label_11
+	j label_12
+	label_11:
+	li $v0 4
+	la $a0 data_17
+	syscall
+	li $v0 10
+	syscall
+	label_12:
 	lw $t0 0($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
-	lw $t0 -12($fp)
+	lw $t0 -20($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	lw $t0 0($fp)
@@ -742,9 +822,9 @@ function_main_at_Main:
 	lw $t0 24($t0)
 	jalr $ra $t0
 	addi $sp $sp 8
-	sw $v0 -20($fp)
-	lw $v0 -20($fp)
-	addi $sp $sp 36
+	sw $v0 -12($fp)
+	lw $v0 -12($fp)
+	addi $sp $sp 56
 	lw $fp 0($sp)
 	addi $sp $sp 4
 	lw $ra 0($sp)
@@ -782,7 +862,7 @@ function_fib_at_Main:
 	sw $v0 -104($fp)
 	lw $t0 -104($fp)
 	sw $t0 -108($fp)
-	label_3:
+	label_13:
 	li $t0 0
 	addi $sp $sp -4
 	sw $t0 0($sp)
@@ -813,25 +893,25 @@ function_fib_at_Main:
 	seq $t0 $t0 $t1
 	sw $t0 -56($fp)
 	lw $t0 -56($fp)
-	bne $t0 $zero label_4
+	bne $t0 $zero label_14
 	lw $t0 -72($fp)
 	lw $t1 -64($fp)
 	seq $t0 $t0 $t1
 	sw $t0 -56($fp)
 	lw $t0 -56($fp)
-	bne $t0 $zero label_5
+	bne $t0 $zero label_15
 	lw $t0 -72($fp)
 	lw $t1 -60($fp)
 	seq $t0 $t0 $t1
 	sw $t0 -56($fp)
 	lw $t0 -56($fp)
-	bne $t0 $zero label_6
+	bne $t0 $zero label_16
 	lw $t0 0($fp)
 	lw $t1 -48($fp)
 	seq $t0 $t0 $t1
 	sw $t0 -52($fp)
-	j label_7
-	label_4:
+	j label_17
+	label_14:
 	lw $t0 0($fp)
 	lw $t1 12($t0)
 	sw $t1 -80($fp)
@@ -842,8 +922,8 @@ function_fib_at_Main:
 	lw $t1 -76($fp)
 	seq $t0 $t0 $t1
 	sw $t0 -52($fp)
-	j label_7
-	label_5:
+	j label_17
+	label_15:
 	lw $t0 0($fp)
 	lw $t1 12($t0)
 	sw $t1 -80($fp)
@@ -856,8 +936,8 @@ function_fib_at_Main:
 	move $a1 $t0
 	jal equal_str
 	sw $v0 -52($fp)
-	j label_7
-	label_6:
+	j label_17
+	label_16:
 	lw $t0 0($fp)
 	lw $t1 12($t0)
 	sw $t1 -80($fp)
@@ -868,8 +948,8 @@ function_fib_at_Main:
 	lw $t1 -76($fp)
 	seq $t0 $t0 $t1
 	sw $t0 -52($fp)
-	j label_7
-	label_7:
+	j label_17
+	label_17:
 	lw $t0 -52($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
@@ -892,9 +972,9 @@ function_fib_at_Main:
 	lw $t1 12($t0)
 	sw $t1 -100($fp)
 	lw $t0 -100($fp)
-	bne $t0 $zero label_8
-	j label_9
-	label_8:
+	bne $t0 $zero label_18
+	j label_19
+	label_18:
 	lw $t0 -124($fp)
 	lw $t1 12($t0)
 	sw $t1 -40($fp)
@@ -941,8 +1021,8 @@ function_fib_at_Main:
 	sw $t0 -116($fp)
 	lw $t0 -108($fp)
 	sw $t0 -124($fp)
-	j label_3
-	label_9:
+	j label_13
+	label_19:
 	lw $v0 -108($fp)
 	addi $sp $sp 116
 	lw $fp 0($sp)
