+++ tests/codegen/hello_world.mips
@@ -23,6 +23,8 @@ data_9: .asciiz "Abort called from class "
 data_10: .asciiz "Substring out of range"
 .word 0
 data_11: .asciiz "Hello, World.\n"
+.word 0
+data_12: .asciiz "RuntimeError: Function call in a void instance"
 
 
 types_table:
@@ -646,19 +648,33 @@ function_main_at_Main:
 	addi $sp $sp -4
 	sw $fp 0($sp)
 	addi $fp $sp 8
-	addi $sp $sp -12
+	addi $sp $sp -16
 	la $t0 data_11
-	sw $t0 -16($fp)
-	lw $t0 -16($fp)
+	sw $t0 -24($fp)
+	lw $t0 -24($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	jal __init_at_String
 	addi $sp $sp 4
-	sw $v0 -12($fp)
+	sw $v0 -20($fp)
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -16($fp)
+	lw $t0 -16($fp)
+	bne $t0 $zero label_3
+	j label_4
+	label_3:
+	li $v0 4
+	la $a0 data_12
+	syscall
+	li $v0 10
+	syscall
+	label_4:
 	lw $t0 0($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
-	lw $t0 -12($fp)
+	lw $t0 -20($fp)
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	lw $t0 0($fp)
@@ -670,9 +686,9 @@ function_main_at_Main:
 	lw $t0 24($t0)
 	jalr $ra $t0
 	addi $sp $sp 8
-	sw $v0 -20($fp)
-	lw $v0 -20($fp)
-	addi $sp $sp 12
+	sw $v0 -12($fp)
+	lw $v0 -12($fp)
+	addi $sp $sp 16
 	lw $fp 0($sp)
 	addi $sp $sp 4
 	lw $ra 0($sp)
