+++ tests/codegen/print-cool.mips
@@ -21,6 +21,22 @@ data_8: .asciiz ""
 data_9: .asciiz "Abort called from class "
 .word 0
 data_10: .asciiz "Substring out of range"
+.word 0
+data_11: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_12: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_13: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_14: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_15: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_16: .asciiz "RuntimeError: Function call in a void instance"
+.word 0
+data_17: .asciiz "\n"
+.word 0
+data_18: .asciiz "RuntimeError: Function call in a void instance"
 
 
 types_table:
@@ -644,20 +660,263 @@ function_main_at_Main:
 	addi $sp $sp -4
 	sw $fp 0($sp)
 	addi $fp $sp 8
-	addi $sp $sp -16
+	addi $sp $sp -92
 	li $t0 1
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	jal __init_at_Int
 	addi $sp $sp 4
-	sw $v0 -24($fp)
+	sw $v0 -100($fp)
 	li $t0 3
 	addi $sp $sp -4
 	sw $t0 0($sp)
 	jal __init_at_Int
 	addi $sp $sp 4
+	sw $v0 -96($fp)
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -88($fp)
+	lw $t0 -88($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	jal __init_at_Bool
+	addi $sp $sp 4
+	sw $v0 -92($fp)
+	lw $t0 -92($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -84($fp)
+	lw $t0 -84($fp)
+	bne $t0 $zero label_3
+	j label_4
+	label_3:
+	li $v0 4
+	la $a0 data_11
+	syscall
+	li $v0 10
+	syscall
+	label_4:
+	lw $t0 -92($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -92($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($sp)
+	addi $sp $sp 4
+	lw $t0 8($t0)
+	lw $t0 4($t0)
+	jalr $ra $t0
+	addi $sp $sp 4
+	sw $v0 -80($fp)
+	lw $t0 -80($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -76($fp)
+	lw $t0 -76($fp)
+	bne $t0 $zero label_5
+	j label_6
+	label_5:
+	li $v0 4
+	la $a0 data_12
+	syscall
+	li $v0 10
+	syscall
+	label_6:
+	lw $t0 -80($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -100($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -96($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -80($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($sp)
+	addi $sp $sp 4
+	lw $t0 8($t0)
+	lw $t0 20($t0)
+	jalr $ra $t0
+	addi $sp $sp 12
+	sw $v0 -72($fp)
+	li $t0 4
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	jal __init_at_Int
+	addi $sp $sp 4
+	sw $v0 -68($fp)
+	li $t0 1
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	jal __init_at_Int
+	addi $sp $sp 4
+	sw $v0 -64($fp)
+	jal __init_at_Object
+	sw $v0 -60($fp)
+	lw $t0 -60($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -56($fp)
+	lw $t0 -56($fp)
+	bne $t0 $zero label_7
+	j label_8
+	label_7:
+	li $v0 4
+	la $a0 data_13
+	syscall
+	li $v0 10
+	syscall
+	label_8:
+	lw $t0 -60($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -60($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($sp)
+	addi $sp $sp 4
+	lw $t0 8($t0)
+	lw $t0 4($t0)
+	jalr $ra $t0
+	addi $sp $sp 4
+	sw $v0 -52($fp)
+	lw $t0 -52($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -48($fp)
+	lw $t0 -48($fp)
+	bne $t0 $zero label_9
+	j label_10
+	label_9:
+	li $v0 4
+	la $a0 data_14
+	syscall
+	li $v0 10
+	syscall
+	label_10:
+	lw $t0 -52($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -68($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -64($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -52($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($sp)
+	addi $sp $sp 4
+	lw $t0 8($t0)
+	lw $t0 20($t0)
+	jalr $ra $t0
+	addi $sp $sp 12
+	sw $v0 -44($fp)
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -40($fp)
+	lw $t0 -40($fp)
+	bne $t0 $zero label_11
+	j label_12
+	label_11:
+	li $v0 4
+	la $a0 data_15
+	syscall
+	li $v0 10
+	syscall
+	label_12:
+	lw $t0 0($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -44($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($sp)
+	addi $sp $sp 4
+	lw $t0 8($t0)
+	lw $t0 24($t0)
+	jalr $ra $t0
+	addi $sp $sp 8
+	sw $v0 -36($fp)
+	lw $t0 -36($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -32($fp)
+	lw $t0 -32($fp)
+	bne $t0 $zero label_13
+	j label_14
+	label_13:
+	li $v0 4
+	la $a0 data_16
+	syscall
+	li $v0 10
+	syscall
+	label_14:
+	lw $t0 -36($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -72($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -36($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($sp)
+	addi $sp $sp 4
+	lw $t0 8($t0)
+	lw $t0 24($t0)
+	jalr $ra $t0
+	addi $sp $sp 8
+	sw $v0 -28($fp)
+	la $t0 data_17
+	sw $t0 -24($fp)
+	lw $t0 -24($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	jal __init_at_String
+	addi $sp $sp 4
 	sw $v0 -20($fp)
-	addi $sp $sp 16
+	lw $t0 0($fp)
+	li $t1 0
+	seq $t0 $t0 $t1
+	sw $t0 -16($fp)
+	lw $t0 -16($fp)
+	bne $t0 $zero label_15
+	j label_16
+	label_15:
+	li $v0 4
+	la $a0 data_18
+	syscall
+	li $v0 10
+	syscall
+	label_16:
+	lw $t0 0($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 -20($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($fp)
+	addi $sp $sp -4
+	sw $t0 0($sp)
+	lw $t0 0($sp)
+	addi $sp $sp 4
+	lw $t0 8($t0)
+	lw $t0 24($t0)
+	jalr $ra $t0
+	addi $sp $sp 8
+	sw $v0 -12($fp)
+	lw $v0 -12($fp)
+	addi $sp $sp 92
 	lw $fp 0($sp)
 	addi $sp $sp 4
 	lw $ra 0($sp)
